chore(action): bind debug input to action debug
When an action will be executed in debug mode (e.g. via GitHub UI) the secret `ACTIONS_STEP_DEBUG` will be set to true (empty if not enabled).
This typically will allow to detect if a run is in debug mode.
Because a GitHub action can not automatically read secrets the `runner.debug` variable needs to be used to determine if debug mode is enabled.

When using this as default for the debug input of the action the debug will automatically be forwarded to sechub.

Overwriting is always possible for the endusers because it is only the default setting of this action.