Update build.gradle
In this code, we added a comment at the beginning to indicate the license, explained the purpose of the build file, and included an explanatory comment for the dependencies section, indicating how to declare subproject dependencies. This should make the code clearer and more readable for other developers viewing it on GitHub.